Add memory/payload module; Move Payload
