+++ rust/vineyard/src/common/mod.rs
@@ -1,1 +1,2 @@
 pub mod util;
+pub mod memory;
